Struct uint::U256

#[repr(C)]

pub struct U256(pub [u64; 4]);

Little-endian large integer type

Methods

impl U256

pub const MAX: U256

MAX: U256 = U256([<u64>::max_value(); 4])

pub fn from_dec_str(value: &str) -> Result<Self, FromDecStrErr>

Convert from a decimal string.

pub fn low_u32(&self) -> u32

Conversion to u32

pub fn low_u64(&self) -> u64

Conversion to u64

pub fn as_u32(&self) -> u32

Conversion to u32 with overflow checking

Panics

Panics if the number is larger than 2^32.

pub fn as_u64(&self) -> u64

Conversion to u64 with overflow checking

Panics

Panics if the number is larger than 2^64.

pub fn as_usize(&self) -> usize

Conversion to usize with overflow checking

Panics

Panics if the number is larger than usize::max_value().

pub fn is_zero(&self) -> bool

Whether this is zero.

pub fn bits(&self) -> usize

Return the least number of bits needed to represent the number

pub fn bit(&self, index: usize) -> bool

Return if specific bit is set.

Panics

Panics if index exceeds the bit width of the number.

pub fn leading_zeros(&self) -> u32

Returns the number of leading zeros in the binary representation of self.

pub fn trailing_zeros(&self) -> u32

Returns the number of leading zeros in the binary representation of self.

pub fn byte(&self, index: usize) -> u8

Return specific byte.

Panics

Panics if index exceeds the byte width of the number.

pub fn to_big_endian(&self, bytes: &mut [u8])

Write to the slice in big-endian format.

pub fn to_little_endian(&self, bytes: &mut [u8])

Write to the slice in little-endian format.

pub fn exp10(n: usize) -> Self

Create 10**n as this type.

Panics

Panics if the result overflows the type.

pub fn zero() -> Self

Zero (additive identity) of this type.

pub fn one() -> Self

One (multiplicative identity) of this type.

pub fn max_value() -> Self

The maximum value which can be inhabited by this type.

pub fn pow(self, expon: Self) -> Self

Fast exponentation by squaring https://en.wikipedia.org/wiki/Exponentiation_by_squaring

Panics

Panics if the result overflows the type.

pub fn overflowing_pow(self, expon: Self) -> (Self, bool)

Fast exponentation by squaring https://en.wikipedia.org/wiki/Exponentiation_by_squaring

pub fn overflowing_add(self, other: U256) -> (U256, bool)

Optimized instructions

pub fn saturating_add(self, other: U256) -> U256

Addition which saturates at the maximum value.

pub fn checked_add(self, other: U256) -> Option<U256>

Checked addition. Returns None if overflow occurred.

pub fn overflowing_sub(self, other: U256) -> (U256, bool)

Subtraction which underflows and returns a flag if it does.

pub fn saturating_sub(self, other: U256) -> U256

Subtraction which saturates at zero.

pub fn checked_sub(self, other: U256) -> Option<U256>

Checked subtraction. Returns None if overflow occurred.

pub fn overflowing_mul(self, other: U256) -> (U256, bool)

Multiply with overflow, returning a flag if it does.

pub fn saturating_mul(self, other: U256) -> U256

Multiplication which saturates at the maximum value..

pub fn checked_mul(self, other: U256) -> Option<U256>

Checked multiplication. Returns None if overflow occurred.

pub fn overflowing_div(self, other: U256) -> (U256, bool)

Division with overflow

pub fn checked_div(self, other: U256) -> Option<U256>

Checked division. Returns None if other == 0.

pub fn overflowing_rem(self, other: U256) -> (U256, bool)

Modulus with overflow.

pub fn checked_rem(self, other: U256) -> Option<U256>

Checked modulus. Returns None if other == 0.

pub fn overflowing_neg(self) -> (U256, bool)

Negation with overflow.

pub fn checked_neg(self) -> Option<U256>

Checked negation. Returns None unless self == 0.

pub fn mul_u32(self, other: u32) -> Self

Deprecated
: Use Mul<u32> instead.

Multiplication by u32

pub fn from_big_endian(slice: &[u8]) -> Self

Converts from big endian representation bytes in memory Can also be used as (&slice).into(), as it is default From slice implementation for U256

pub fn from_little_endian(slice: &[u8]) -> Self

Converts from little endian representation bytes in memory